+++ CSE/IO/cPreferences.vb
@@ -19,7 +19,7 @@ Public Class cPreferences
 
     Protected Overrides Function HeaderOverlay() As JObject
         Return JObject.Parse(<json>{
-                "Title": "vecto-cse PREFERENCES",
+                "Title": "VECTO-Air Drag PREFERENCES",
                 "FileVersion":  "1.0.0",
            }</json>.Value)
     End Function
@@ -42,7 +42,7 @@ Public Class cPreferences
     Public Shared Function JSchemaStr(Optional ByVal isStrictBody As Boolean = False) As String
         Dim allowAdditionalProps_str As String = (Not isStrictBody).ToString.ToLower
         Return <json>{
-            "title": "Schema for vecto-cse PREFERENCES",
+            "title": "Schema for VECTO-Air Drag PREFERENCES",
             "type": "object", "additionalProperties": <%= allowAdditionalProps_str %>, 
             "required": true,
             "properties": {